added submenu scenes, reworked script to fit
there is probably going to be a conflict with adam's branch bc i did a similar change to the submenu script there but just use this one and it should be fine
