--- Assets/General/Scenes/BuildScenes/SettingsMenuScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c7ec97aadc2c6a54898e142d8157a8be
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
