+++ Assets/General/Scenes/BuildScenes/Submenus/QuitConfirm.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5306edfc976c48240aaaddf7d87afef3
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
