+++ Assets/General/Scenes/BuildScenes/Submenus/SettingsMenuScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c7ec97aadc2c6a54898e142d8157a8be
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
