+++ ProjectSettings/EditorBuildSettings.asset
@@ -18,6 +18,15 @@ EditorBuildSettings:
     path: Assets/General/Scenes/BuildScenes/EndScreen.unity
     guid: acf64113057aadd42bb7a95c6ec927b9
   - enabled: 1
-    path: Assets/General/Scenes/BuildScenes/SettingsMenuScene.unity
+    path: Assets/General/Scenes/BuildScenes/Submenus/SettingsMenuScene.unity
     guid: c7ec97aadc2c6a54898e142d8157a8be
+  - enabled: 1
+    path: Assets/General/Scenes/BuildScenes/Submenus/OptionsSubMenu.unity
+    guid: b1d9b2b0565619b48b5e90b986b317ec
+  - enabled: 1
+    path: Assets/General/Scenes/BuildScenes/Submenus/Controls.unity
+    guid: 7123a90ec4cd494449974f84191a92c2
+  - enabled: 1
+    path: Assets/General/Scenes/BuildScenes/Submenus/QuitConfirm.unity
+    guid: 5306edfc976c48240aaaddf7d87afef3
   m_configObjects: {}
